Move view mode types default to useViewMode hook.
